refactor: move bounding box tests into a new file
