+++ tests/test_bbox.py
@@ -1,0 +1,82 @@
+import pytest
+from affine import Affine
+
+from odc.geo import BoundingBox, wh_
+from odc.geo.geom import bbox_intersection, bbox_union
+from odc.geo.testutils import epsg3857
+
+
+def test_boundingbox():
+    bb = BoundingBox(0, 3, 2, 4)
+    assert bb.width == 2
+    assert bb.height == 1
+    assert bb.width == bb.span_x
+    assert bb.height == bb.span_y
+    assert bb.shape == (bb.height, bb.width)
+    assert bb.crs is None
+    assert bb.bbox == (0, 3, 2, 4)
+    assert "crs=" not in str(bb)
+    assert "crs=" not in repr(bb)
+
+    bb = BoundingBox(0, 3, 2.1, 4, "epsg:4326")
+    assert bb.width == 2
+    assert bb.height == 1
+    assert bb.span_x == 2.1
+    assert bb.width != bb.span_x
+    assert bb.height == bb.span_y
+    assert bb.crs.epsg == 4326
+    assert "crs=" in str(bb)
+    assert "crs=" in repr(bb)
+
+    assert BoundingBox.from_xy(bb.range_x, bb.range_y, bb.crs) == bb
+
+    assert BoundingBox.from_xy((1, 2), (10, 20)) == (1, 10, 2, 20)
+    assert BoundingBox.from_xy((2, 1), (20, 10)) == (1, 10, 2, 20)
+    assert BoundingBox.from_points((1, 11), (2, 22)) == (1, 11, 2, 22)
+    assert BoundingBox.from_points((1, 22), (2, 11)) == (1, 11, 2, 22)
+
+    assert BoundingBox(1, 3, 10, 20).buffered(1, 3) == (0, 0, 11, 23)
+    assert BoundingBox(1, 3, 10, 20).buffered(1) == (0, 2, 11, 21)
+
+    assert BoundingBox.from_transform(
+        wh_(10, 12), Affine.translation(1, 3), "epsg:3857"
+    ) == (1, 3, 11, 15)
+    assert (
+        BoundingBox.from_transform(
+            wh_(10, 12), Affine.translation(1, 3), "epsg:3857"
+        ).crs
+        == "epsg:3857"
+    )
+
+    assert bb.polygon.boundingbox == bb
+
+
+@pytest.mark.parametrize("crs", [None, "epsg:4326", epsg3857])
+def test_bbox_union(crs):
+    b1 = BoundingBox(0, 1, 10, 20, crs)
+    b2 = BoundingBox(5, 6, 11, 22, crs)
+
+    assert bbox_union([b1]) == b1
+    assert bbox_union([b2]) == b2
+
+    bb = bbox_union(iter([b1, b2]))
+    assert bb == BoundingBox(0, 1, 11, 22, crs)
+
+    bb = bbox_union(iter([b2, b1] * 10))
+    assert bb == BoundingBox(0, 1, 11, 22, crs)
+
+
+@pytest.mark.parametrize(
+    "crss",
+    [
+        (None, "epsg:4326", epsg3857),
+        ("epsg:4326", "epsg:3857"),
+        (*["epsg:4326"] * 4, "epsg:3857"),
+    ],
+)
+def test_bbox_crs_mismatch(crss):
+    with pytest.raises(ValueError):
+        _ = bbox_union(BoundingBox(0, 0, 1, 1, crs) for crs in crss)
+
+    with pytest.raises(ValueError):
+        _ = bbox_intersection(BoundingBox(0, 0, 1, 1, crs) for crs in crss)
